requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.0.0
WTForms==3.0.1
python-barcode==0.14.0
Pillow==9.3.0
cryptography==38.0.4
Werkzeug==2.1.2